switched verification to a status page
